+++ golang/lib/services/container_config_factory.go
@@ -1,5 +1,9 @@
 package services
 
 type ContainerConfigFactory interface {
-	Create(containerIpAddr string) *ContainerConfig
+	// TODO DOCS LINK
+	GetCreationConfig(containerIpAddr string) *ContainerCreationConfig
+
+	// TODO DOCS Link
+	GetRunConfig(containerIpAddr string, generatedFileFilepaths map[string]string) *ContainerRunConfig
 }